2005-07-13 Lillian Angel <langel@redhat.com>         * javax/swing/plaf/basic/BasicLookAndFeel.java: Changed rowHeight         to the right height. This should be fixed later to something more         reasonable, as opposed to a 'magic' number.         * javax/swing/plaf/basic/BasicTreeUI.java         (paintLeaf): added in code to paint icons properly         (paintNonLeaf): same as above         * javax/swing/plaf/metal/MetalLookAndFeel.java: Changed default         icons to MetalIconFactory icons         * javax/swing/tree/DefaultTreeCellRenderer.java         (getTreeCellRendererComponent): Changed to paint icons separately from         Cell.

+++ javax/swing/plaf/basic/BasicLookAndFeel.java
@@ -1041,7 +1041,7 @@ public abstract class BasicLookAndFeel extends LookAndFeel
       "Tree.hash", new ColorUIResource(Color.gray),
       "Tree.leftChildIndent", new Integer(7),
       "Tree.rightChildIndent", new Integer(13),
-      "Tree.rowHeight", new Integer(19),
+      "Tree.rowHeight", new Integer(20), // FIXME
       "Tree.scrollsOnExpand", Boolean.TRUE,
       "Tree.selectionBackground", new ColorUIResource(Color.black),
       "Tree.nonSelectionBackground", new ColorUIResource(light),
